[refactor/#46] 홈화면 섹션별 위젯 분리
