+++ src/entities/recipe/ui/BaseFoodCard.tsx
@@ -68,10 +68,7 @@ export function BaseFoodCard({ recipe, variant, onPress, likeButton }: BaseFoodC
   );
 
   return (
-    <Pressable
-      className="overflow-hidden rounded-hero bg-surface-card shadow-sm"
-      onPress={onPress}
-    >
+    <Pressable className="overflow-hidden rounded-hero bg-surface-card shadow-sm" onPress={onPress}>
       <View className={imageHeightClassName} onLayout={handleBannerLayout}>
         <Image source={{ uri: recipe.thumbnail }} className="h-full w-full" resizeMode="cover" />
         {bannerOverlay}
