+++ src/widgets/fridge-status-bar/index.ts
@@ -1,0 +1,1 @@
+export { FridgeStatusBar } from "./ui/FridgeStatusBar";
